4: Footer now positioned via screen size changes
When the screen size changes, the footer is placed always such that it is at the
bottom, even with changes to the fonts.

NOTE:  This is done with raw pixels, so the footer offset must be changed when
any changes to the font sizes are implemented.  Otherwise the footer will not be
positioned correctly.